refactor: 将 `StepManager` 从 `steps.py` 移动到 `output.py` 并更新了相关模块。
